requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
idna==2.8
python-craigslist==1.0.8
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.8
